requirements.txt
ultralytics>=8.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
matplotlib>=3.7.0